requests and first dialogs

+++ Windows.py
@@ -1,12 +1,14 @@
-from qgis.PyQt.QtWidgets import QAction, QDesktopWidget, QMainWindow, QHBoxLayout, QComboBox, QVBoxLayout, QWidget, QFileDialog, QGridLayout, QPushButton, QLabel, QLineEdit, QDialog
+from qgis.PyQt.QtWidgets import QDesktopWidget, QMainWindow, QHBoxLayout, QComboBox, QVBoxLayout, QWidget, QGridLayout, QPushButton, QLabel, QLineEdit, QDialog
 from qgis.PyQt.QtGui import QIcon
-from qgis.core import Qgis
+from qgis.core import Qgis, QgsPointXY
 from qgis.gui import QgsMapTool
 
 from .EquiView360 import GLWidget
 from .DBHandler import connector
+from .Helpers import MapManager
 
 import json
+import os
 
 class ConnectionDialog(QDialog):
     def __init__(self, parent=None):
@@ -34,7 +36,6 @@ class ConnectionDialog(QDialog):
         button_layout.addWidget(button_connect)
         button_layout.addWidget(button_cancel)
 
-
         grid = QGridLayout()
         grid.addWidget(label_host, 0, 0)
         grid.addWidget(self.lineEdit_host, 0, 1)
@@ -52,8 +53,8 @@ class ConnectionDialog(QDialog):
         grid.addWidget(self.lineEdit_table, 6, 1)
         grid.addLayout(button_layout, 7, 0, 1, 2)
 
-
         self.setLayout(grid)
+        self.setFixedWidth(400)
 
         button_connect.clicked.connect(self.accept)
         button_cancel.clicked.connect(self.reject)
@@ -71,11 +72,6 @@ class ConnectionDialog(QDialog):
             pass
 
     def get_connection(self):
-        """
-        Get the connection parameters entered by the user.
-        Returns:
-            A tuple containing the host, port, database, username, and password.
-        """
         host = self.lineEdit_host.text()
         port = self.lineEdit_port.text()
         database = self.lineEdit_database.text()
@@ -92,7 +88,6 @@ class ConnectionDialog(QDialog):
         username = self.lineEdit_username.text()
         schema = self.lineEdit_schema.text()
         table = self.lineEdit_table.text()
-
         connection_params = {
             "host": host,
             "port": port,
@@ -101,20 +96,29 @@ class ConnectionDialog(QDialog):
             "schema": schema,
             "table": table,
         }
-        with open("connection_params.json", "w") as f:
-            json.dump(connection_params, f)
+        if os.path.exists("connection_params.json"):
+            with open("connection_params.json", "r") as f:
+                connection_params = json.load(f)
+            connection_params["host"] = host
+            connection_params["port"] = port
+            connection_params["database"] = database
+            connection_params["username"] = username
+            connection_params["schema"] = schema
+            connection_params["table"] = table
+        else : 
+            with open("connection_params.json", "w") as f:
+                json.dump(connection_params, f)
+
     def accept(self):
         if not all([self.lineEdit_host.text(), self.lineEdit_port.text(), self.lineEdit_database.text(), self.lineEdit_username.text(), self.lineEdit_password.text(), self.lineEdit_schema.text(), self.lineEdit_table.text()]):
-            # Display an error message or do something else to indicate to the user that all fields are required
             pass
         else:
             super().accept()
 
 class MainWindow(QMainWindow):
-    def __init__(self, iface, url):
+    def __init__(self, iface, url, map_manager):
         super().__init__()
-
-        # Create the horizontal layout for the list widgets
+        self.map_manager = map_manager
         horizontalLayout = QHBoxLayout()
         comboBox1 = QComboBox()
         comboBox2 = QComboBox()
@@ -123,37 +127,48 @@ class MainWindow(QMainWindow):
         horizontalLayout.addWidget(comboBox2)
         horizontalLayout.addWidget(comboBox3)
 
-        # Create the vertical layout for the GLWidget and list widgets
         verticalLayout = QVBoxLayout()
         self.gl_widget = GLWidget(self, iface, url)
         verticalLayout.addWidget(self.gl_widget)
         verticalLayout.addLayout(horizontalLayout)
 
-        # Set the main window's central widget and layout
         centralWidget = QWidget()
         centralWidget.setLayout(verticalLayout)
         self.setCentralWidget(centralWidget)
 
-        # Set the window properties
         self.setWindowTitle("Equirectangular 360° Viewer")
         self.setWindowIcon(QIcon("icon.png"))
-        self.setGeometry(0, 0, 1080, 720)
+        self.resize(1080,720)
+    
+    def closeEvent(self, event):
+        self.map_manager.remove_all_points_from_map()
+        super().closeEvent(event)
+
 
 class PointTool(QgsMapTool):  
 
-    def __init__(self, canvas, iface, cursor):
+    def __init__(self, canvas, iface, cursor, geom, yaw, link, schema, table):
         QgsMapTool.__init__(self, canvas)
         self.iface = iface
         self.canvas = canvas
         self.cursor = cursor
+        self.schema = schema
+        self.table = table
+        self.link = link
+        self.geom = geom
+        self.yaw = yaw
+        self.direction = None
 
     def canvasPressEvent(self, event):
-        t = event.pos().x()
-        w = event.pos().y()
-        point = self.canvas.getCoordinateTransform().toMapCoordinates(t, w)
-        url = connector(point.x(), point.y(), self.cursor)
-        if url != 0 :
-            self.dlg = MainWindow(self.iface, url)
+        x = event.pos().x()
+        y = event.pos().y()
+        point = self.canvas.getCoordinateTransform().toMapCoordinates(x, y)
+        url, direction = connector(point.x(), point.y(), self.cursor, self.geom, self.schema, self.table, self.link, self.yaw)
+
+        if url != 0 and direction is not None :
+            map_manager = MapManager(self.canvas)
+            map_manager.add_point_to_map(point, float(direction))
+            self.dlg = MainWindow(self.iface, url, map_manager)
             screen = QDesktopWidget().screenGeometry()
             size = self.dlg.geometry()
             x = (screen.width() - size.width()) / 2
@@ -161,44 +176,90 @@ class PointTool(QgsMapTool):
             self.dlg.move(x, y)
             self.dlg.show()
             self.iface.mapCanvas().unsetMapTool(self)
+
         else : 
             self.iface.messageBar().pushMessage("No image for this coordinates", level=Qgis.Info)
+    
 
 class ColumnSelectionDialog(QDialog):
     def __init__(self, columns, parent=None):
         super().__init__(parent)
-        
-        # Create a label and combo box for each column
+        self.geom_label = QLabel("Geometry column:")
+        self.geom_combo = QComboBox()
         self.yaw_label = QLabel("Direction column:")
         self.yaw_combo = QComboBox()
         self.link_label = QLabel("Link column:")
         self.link_combo = QComboBox()
         self.date_label = QLabel("Date column:")
         self.date_combo = QComboBox()
-        self.fov_label = QLabel("FOV column:")
-        self.fov_combo = QComboBox()
-        
-        # Populate the combo boxes with the available columns
+
+        self.geom_combo.addItems(columns)
         self.yaw_combo.addItems(columns)
         self.link_combo.addItems(columns)
         self.date_combo.addItems(columns)
-        self.fov_combo.addItems(columns)
         
-        # Create a button to accept the selected columns
+
+        try : 
+            with open("connection_params.json", "r") as f:
+                connection_params = json.load(f)
+                default_geom = connection_params["geom"]
+                default_yaw = connection_params["yaw"]
+                default_link = connection_params["link"]
+                default_date = connection_params["date"]
+                default_geom_index = self.date_combo.findText(default_geom)
+                self.geom_combo.setCurrentIndex(default_geom_index)
+                default_yaw_index = self.yaw_combo.findText(default_yaw)
+                self.yaw_combo.setCurrentIndex(default_yaw_index)
+                default_link_index = self.link_combo.findText(default_link)
+                self.link_combo.setCurrentIndex(default_link_index)
+                default_date_index = self.date_combo.findText(default_date)
+                self.date_combo.setCurrentIndex(default_date_index)
+        except Exception : 
+            pass
+
         self.ok_button = QPushButton("OK")
         self.ok_button.clicked.connect(self.accept)
+        self.ok_button.clicked.connect(self.save_data)
         
-        # Create a vertical layout to hold the widgets
         layout = QVBoxLayout()
+        layout.addWidget(self.geom_label)
+        layout.addWidget(self.geom_combo)
         layout.addWidget(self.yaw_label)
         layout.addWidget(self.yaw_combo)
         layout.addWidget(self.link_label)
         layout.addWidget(self.link_combo)
         layout.addWidget(self.date_label)
         layout.addWidget(self.date_combo)
-        layout.addWidget(self.fov_label)
-        layout.addWidget(self.fov_combo)
         layout.addWidget(self.ok_button)
         
-        # Set the layout of the dialog
         self.setLayout(layout)
+        self.setFixedWidth(400)
+
+    def save_data(self):
+        geom = self.geom_combo.currentText()
+        yaw = self.yaw_combo.currentText()
+        link = self.link_combo.currentText()
+        date = self.date_combo.currentText()
+        
+
+        connection_params = {
+            "geom": geom,
+            "yaw": yaw,
+            "link": link,
+            "date": date,
+        }
+        with open("connection_params.json", "r") as f:
+            connection_params = json.load(f)
+            connection_params["geom"] = geom
+            connection_params["yaw"] = yaw
+            connection_params["link"] = link
+            connection_params["date"] = date
+        with open("connection_params.json", "w") as f:
+            json.dump(connection_params, f)
+
+    def get_columns(self):
+        geom = self.geom_combo.currentText()
+        yaw = self.yaw_combo.currentText()
+        link = self.link_combo.currentText()
+        date = self.date_combo.currentText()
+        return geom, yaw, link, date
